Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -4,6 +4,6 @@
 int main() {
     float a,b;
     scanf("%f %f",&a,&b);
-    int pro=a*b;
+    float pro=a*b;
     printf("Product: %.2f",pro);
 }
